workload: move out of testutils
Some upcoming work will depend on the workload interface from outside
test code, which is disallowed by the linter if it's in testutils. Plus,
it seems to generally useful enough to graduate.

Release note: None
