+++ pkg/cmd/workload/run.go
@@ -38,10 +38,10 @@ import (
 	"github.com/spf13/pflag"
 	"github.com/tylertreat/hdrhistogram-writer"
 
-	"github.com/cockroachdb/cockroach/pkg/testutils/workload"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/cockroachdb/cockroach/pkg/util/syncutil"
 	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
+	"github.com/cockroachdb/cockroach/pkg/workload"
 )
 
 const crdbDefaultURI = `postgres://root@localhost:26257?sslmode=disable`
