+++ pkg/workload/tpcc/tpcc.go
@@ -20,7 +20,7 @@ import (
 	gosql "database/sql"
 	"sync/atomic"
 
-	"github.com/cockroachdb/cockroach/pkg/testutils/workload"
+	"github.com/cockroachdb/cockroach/pkg/workload"
 	"github.com/spf13/pflag"
 )
